chore: move logic to llm_utils
